# Convert remaining `Node.get*` methods to use paths