--- a/packages/slate-react/src/components/content.js
+++ b/packages/slate-react/src/components/content.js
@@ -461,17 +461,9 @@ class Content extends React.Component {
     const childrenDecorations = getChildrenDecorations(document, decs)
 
     const children = document.nodes.toArray().map((child, i) => {
-      const containedInSelection =
-        !!indexes && indexes.start < i && i < indexes.end - 1
-      const onEdgeOfSelection =
-        !!indexes && (i === indexes.start || i === indexes.end - 1)
+      const isSelected = !!indexes && indexes.start <= i && i < indexes.end
 
-      return this.renderNode(
-        child,
-        { onEdgeOfSelection, containedInSelection },
-        childrenDecorations[i],
-        i
-      )
+      return this.renderNode(child, isSelected, childrenDecorations[i])
     })
 
     const style = {
@@ -530,17 +522,11 @@ class Content extends React.Component {
    * @return {Element}
    */
 
-  renderNode = (
-    child,
-    { onEdgeOfSelection, containedInSelection },
-    decorations,
-    index
-  ) => {
+  renderNode = (child, isSelected, decorations) => {
     const { editor, readOnly } = this.props
     const { value } = editor
     const { document, selection } = value
     const { isFocused } = selection
-    const isSelected = onEdgeOfSelection || containedInSelection
 
     return (
       <Node
@@ -548,9 +534,6 @@ class Content extends React.Component {
         editor={editor}
         decorations={decorations}
         isSelected={isSelected}
-        onEdgeOfSelection={onEdgeOfSelection}
-        containedInSelection={containedInSelection}
-        path={onEdgeOfSelection && [index]}
         isFocused={isFocused && isSelected}
         key={child.key}
         node={child}