--- a/packages/slate-react/src/components/node.js
+++ b/packages/slate-react/src/components/node.js
@@ -124,6 +124,9 @@ class Node extends React.Component {
       editor,
       isSelected,
       isFocused,
+      containedInSelection,
+      onEdgeOfSelection,
+      path,
       node,
       decorations,
       parent,
@@ -131,16 +134,27 @@ class Node extends React.Component {
     } = this.props
     const { value } = editor
     const { selection } = value
-    const indexes = node.getSelectionIndexes(selection, isSelected)
+    // Only do this calculation if we need to
+    const indexes =
+      onEdgeOfSelection && node.getSelectionIndexes(selection, path)
     const decs = decorations.concat(node.getDecorations(editor))
     const childrenDecorations = getChildrenDecorations(node, decs)
     const children = []
 
     node.nodes.forEach((child, i) => {
-      const isChildSelected = !!indexes && indexes.start <= i && i < indexes.end
+      const containedInSelection =
+        containedInSelection ||
+        (!!indexes && indexes.start < i && i < indexes.end - 1)
+      const onEdgeOfSelection =
+        !!indexes && (i === indexes.start || i === indexes.end - 1)
 
       children.push(
-        this.renderNode(child, isChildSelected, childrenDecorations[i])
+        this.renderNode(
+          child,
+          { onEdgeOfSelection, containedInSelection },
+          childrenDecorations[i],
+          i
+        )
       )
     })
 
@@ -197,6 +211,9 @@ class Node extends React.Component {
         decorations={decorations}
         editor={editor}
         isSelected={isSelected}
+        containedInSelection={containedInSelection}
+        onEdgeOfSelection={onEdgeOfSelection}
+        path={onEdgeOfSelection && path.concat(index)}
         isFocused={isFocused && isSelected}
         key={child.key}
         node={child}