--- a/packages/slate/test/models/node/get-active-marks-at-range.js/same-mark-across-blocks.js
+++ b/packages/slate/test/models/node/get-active-marks-at-range.js/same-mark-across-blocks.js
@@ -1,0 +1,34 @@
+/** @jsx h */
+
+import h from '../../../helpers/h'
+import { Set } from 'immutable'
+import { Mark } from 'slate'
+
+export const input = (
+  <value>
+    <document>
+      <paragraph>
+        wo<anchor />
+        <mark type="a">rd</mark>
+      </paragraph>
+      <paragraph>
+        <paragraph>
+          <mark type="a">middle</mark>
+        </paragraph>
+        <paragraph />
+      </paragraph>
+      <paragraph>
+        <mark type="a">
+          an<focus />other
+        </mark>
+        <mark type="b">unselected marked text</mark>
+      </paragraph>
+    </document>
+  </value>
+)
+
+export default function({ document, selection }) {
+  return document.getActiveMarksAtRange(selection)
+}
+
+export const output = Set.of(Mark.create('a'))