--- a/packages/slate/test/models/node/get-closest/top-level.js
+++ b/packages/slate/test/models/node/get-closest/top-level.js
@@ -1,0 +1,21 @@
+/** @jsx h */
+
+import h from '../../../helpers/h'
+import { PathUtils } from 'slate'
+
+export const input = (
+  <value>
+    <document>
+      <paragraph>one</paragraph>
+      <paragraph>two</paragraph>
+      <paragraph>three</paragraph>
+      <paragraph>four</paragraph>
+    </document>
+  </value>
+)
+
+export default function({ document, selection }) {
+  return document.getClosestBlock(PathUtils.create([1]))
+}
+
+export const output = null