--- a/packages/slate/test/models/node/get-next-block/by-key.js
+++ b/packages/slate/test/models/node/get-next-block/by-key.js
@@ -1,0 +1,22 @@
+/** @jsx h */
+
+import h from '../../../helpers/h'
+
+export const input = (
+  <value>
+    <document>
+      <paragraph>one</paragraph>
+      <paragraph>
+        <anchor />two<focus />
+      </paragraph>
+      <paragraph>three</paragraph>
+      <paragraph>four</paragraph>
+    </document>
+  </value>
+)
+
+export default function({ document, selection }) {
+  return document.getNextBlock(selection.end.key)
+}
+
+export const output = <paragraph>three</paragraph>