--- a/packages/slate/test/models/node/get-next-node/next-node-is-ancestor-sibling-with-nested-blocks.js
+++ b/packages/slate/test/models/node/get-next-node/next-node-is-ancestor-sibling-with-nested-blocks.js
@@ -1,0 +1,28 @@
+/** @jsx h */
+
+import h from '../../../helpers/h'
+
+export const input = (
+  <value>
+    <document>
+      <paragraph>one</paragraph>
+      <paragraph>
+        <anchor />two<focus />
+      </paragraph>
+      <paragraph>
+        <paragraph>three</paragraph>
+      </paragraph>
+      <paragraph>four</paragraph>
+    </document>
+  </value>
+)
+
+export default function({ document, selection }) {
+  return document.getNextNode(selection.end.path)
+}
+
+export const output = (
+  <paragraph>
+    <paragraph>three</paragraph>
+  </paragraph>
+)