--- a/packages/slate/test/models/node/get-previous-block/prev-block-is-ancestor-sibling.js
+++ b/packages/slate/test/models/node/get-previous-block/prev-block-is-ancestor-sibling.js
@@ -1,0 +1,22 @@
+/** @jsx h */
+
+import h from '../../../helpers/h'
+
+export const input = (
+  <value>
+    <document>
+      <paragraph>one</paragraph>
+      <paragraph>two</paragraph>
+      <paragraph>
+        <anchor />three<focus />
+      </paragraph>
+      <paragraph>four</paragraph>
+    </document>
+  </value>
+)
+
+export default function({ document, selection }) {
+  return document.getPreviousBlock(selection.end.path)
+}
+
+export const output = <paragraph>two</paragraph>