--- a/packages/slate/test/models/node/get-previous-block/prev-block-is-sibling-descendent.js
+++ b/packages/slate/test/models/node/get-previous-block/prev-block-is-sibling-descendent.js
@@ -1,0 +1,24 @@
+/** @jsx h */
+
+import h from '../../../helpers/h'
+
+export const input = (
+  <value>
+    <document>
+      <paragraph>one</paragraph>
+      <paragraph>
+        <anchor />two<focus />
+      </paragraph>
+      <paragraph>
+        <paragraph>three</paragraph>
+      </paragraph>
+      <paragraph>four</paragraph>
+    </document>
+  </value>
+)
+
+export default function({ document, selection }) {
+  return document.getPreviousBlock(selection.end.path)
+}
+
+export const output = <paragraph>one</paragraph>