… packages/slate/test/models/node/get-selection-indexes/across-blocks-from-nested-node.js
… packages/slate/test/models/node/get-selection-indexes/in-single-block-from-middle-nested-node.js
… packages/slate/test/models/node/get-selection-indexes/in-single-block-from-parent-node.js
… packages/slate/test/models/node/get-selection-indexes/in-single-block.js