#### Is this adding or improving a _feature_ or fixing a _bug_?

Improvement.

Opening this PR right now to get discussion on the basic approach (see below).

#### What's the new behavior?

See #2000 for background.

#### How does this change work?

The difficulty using `path` rather than `key` is that there is no `node.path` equivalent for `node.key`.

There are at least 4 potential approaches (some or all could be combined):
1. Store some sort of pointer / reference to `path` on the node itself, and update all of them whenever a change is made that might affect some paths. This would make the rest of the changes easier, just switch to use `node.path` where we use `node.key`.

2. Store some sort of pointer / reference to `path` as a prop when rendering in the editor to make `getSelectionIndexes` easier to calculate at `render` time. 

(See https://github.com/ianstormtaylor/slate/issues/2413 for related discussion).

3. Create a `getPathForKey(key)` that is more efficient than `getKeysToPathsTable` for when we just need the path for a particular node (we can bail early). This would then allow us to call `getPathForKey(node.key)` where we use `node.key`. We could cache this to some extent, but probably calling it more than once or twice is just as expensive as `getKeysToPathsTable`.

4. Create methods that are similar to current `getFirstText`, `getNextBlock` etc, but which return the node and path. This allows some of the more complex methods such as `getActiveMarksAtRange` to iterate internally and still continue to use paths to call various getters.

This PR has three parts:
1. Adds tests
2. Some work towards (4) (see e.g. `getNextTextAndPath`, which can be used to improve the perf of `getActiveMarksAtRange`)
3. A promising approach for (2)

Opening now so we can discuss the approaches / tradeoffs.

If we don't like them, I can just add the tests in a separate PR (fwiw, some of them are failing against `master`, but I'm pretty sure I've captured the expected / desired behavior, and so those might represent legitimate bugs).

#### Have you checked that...?

* [x] The new code matches the existing patterns and styles.
* [x] The tests pass with `yarn test`.
* [x] The linter passes with `yarn lint`. (Fix errors with `yarn prettier`.)
* [ ] The relevant examples still work. (Run examples with `yarn watch`.)

#### Does this fix any issues or need any specific reviewers?

Fixes: #2000
Reviewers: @ianstormtaylor @alanctkc