@jtadmor 2019-01-23
(Unfortunately it seems that many of the `changes` and `models` benchmarks no longer work post `0.42.0` release that changes `value.change` -> `editor.change`.

I will try to separately take some time to refactor those so we can use them going forward.