@jtadmor 2019-02-07
So here is what we have so far:

```
benchmarks
    changes
      ✔ add-mark: 61.24 -> 76.74 ops/sec (25.30% faster)
      ★ delete-backward:
            user: 88.62 -> 117.55 ops/sec (32.64% faster)
            real: 99.63 -> 138.03 ops/sec (38.54% faster)
      ★ delete-forward:
            user: 94.62 -> 129.18 ops/sec (36.53% faster)
            real: 106.99 -> 147.76 ops/sec (38.10% faster)
      ✖ insert-node-by-key:
            user: NaN -> NaN ops/sec (NaN% slower)
            real: NaN -> NaN ops/sec (NaN% slower)
      ✔ insert-text-by-key: 212.15 -> 247.90 ops/sec (16.85% faster)
      ★ insert-text:
            user: 71.20 -> 95.65 ops/sec (34.34% faster)
            real: 79.38 -> 113.80 ops/sec (43.35% faster)
      ★ normalize:
            user: 88.52 -> 134.62 ops/sec (52.08% faster)
            real: 97.56 -> 160.76 ops/sec (64.78% faster)
      ✔ remove-node-by-key: 89.72 -> 107.05 ops/sec (19.31% faster)
      ★ set-blocks:
            user: 143.84 -> 192.14 ops/sec (33.57% faster)
            real: 155.11 -> 212.33 ops/sec (36.90% faster)
      ★ split-block:
            user: 42.77 -> 56.03 ops/sec (30.99% faster)
            real: 47.44 -> 64.22 ops/sec (35.36% faster)
    models
      ✔ from-json-big: 13.20 -> 13.22 ops/sec (0.15% faster)
      ✔ from-json: 673.32 -> 624.56 ops/sec (8.00% slower)
      ★ get-active-marks-at-range:
            user: 2425.19 -> 7441.32 ops/sec (206.83% faster 🙌 )
            real: 2882.27 -> 11225.81 ops/sec (289.48% faster 🙌 )
      ★ get-ancestors:
            user: 2207.39 -> 3887.20 ops/sec (76.10% faster)
            real: 2483.08 -> 4408.68 ops/sec (77.55% faster)
      ✔ get-blocks: 12054.74 -> 10864.26 ops/sec (11.00% slower)
      ✔ get-common-ancestor: 5002.48 -> 4502.59 ops/sec (11.00% slower)
      ★ get-descendant:
            user: 3169.75 -> 4791.45 ops/sec (51.16% faster)
            real: 3583.80 -> 5295.84 ops/sec (47.77% faster)
      ✔ get-furthest-ancestor: 5130.20 -> 4837.64 ops/sec (6.00% slower)
      ✔ get-inlines: 13828.54 -> 11670.79 ops/sec (18.00% slower)
      ★ get-leaf-blocks-at-range:
            user: 723.44 -> 1938.30 ops/sec (167.93% faster 🙌 )
            real: 989.10 -> 3374.90 ops/sec (241.21% faster 🙌 )
      ★ get-leaf-inlines-at-range:
            user: 664.17 -> 1024.20 ops/sec (54.21% faster)
            real: 797.17 -> 1314.38 ops/sec (64.88% faster)
      ✔ get-leaves: 1588404.21 -> 1323132.37 ops/sec (20.00% slower)
      ✔ get-marks-at-range: 816.86 -> 984.49 ops/sec (20.52% faster)
      ✔ get-marks: 3263.45 -> 3045.23 ops/sec (7.00% slower)
      ★ get-parent:
            user: 3054.65 -> 4315.50 ops/sec (41.28% faster)
            real: 3347.41 -> 4733.00 ops/sec (41.39% faster)
      ✔ get-path: 4770.06 -> 4701.75 ops/sec (1.00% slower)
      ★ get-texts-at-range:
            user: 1903.06 -> 4531.84 ops/sec (138.13% faster 🙌 )
            real: 2155.01 -> 5494.71 ops/sec (154.97% faster 🙌 )
      ✔ get-texts: 12039.79 -> 11066.59 ops/sec (9.00% slower)
      ★ has-node-multiple:
            user: 2176.28 -> 3666.96 ops/sec (68.50% faster)
            real: 2410.70 -> 4024.98 ops/sec (66.96% faster)
      ★ has-node:
            user: 3087.33 -> 4416.95 ops/sec (43.07% faster)
            real: 3392.67 -> 4788.70 ops/sec (41.15% faster)
      ✔ to-json: 2920.13 -> 2807.39 ops/sec (4.00% slower)
    html-serializer
      ✔ deserialize: 153.02 -> 158.60 ops/sec (3.65% faster)
      ✔ serialize: 692.19 -> 694.58 ops/sec (0.35% faster)
    plain-serializer
      ✔ deserialize: 3221.63 -> 3915.51 ops/sec (21.54% faster)
      ✔ serialize: 3908.12 -> 4062.15 ops/sec (3.94% faster)
    rendering
      ✔ decoration: 55.34 -> 62.83 ops/sec (13.53% faster)
      ✔ large-with-selection: 34.23 -> 39.28 ops/sec (14.73% faster)
      ✔ normal: 67.86 -> 73.46 ops/sec (8.25% faster)
```