@jtadmor 2019-02-07
Some notes / thoughts:

- Low level benchmarks seem much improved. This is good. However, I'm not convinced this necessarily represents dramatic improvements in the places where perf is hurting (normalization, big changes). I may work on a separate PR to try to get some better / more complex benchmarks in, and encourage others to do likewise while this PR is discussed.
- Passing `path` instead of `key` to a lot of these methods will likely break a lot of the current memoization. We are using `WeakMap` but unless the same exact `List` is passed in, object equality will not hold and it will fail. We should consider changing memoization to coerce any `List` args with a `.join(.)`. This will slow down memoization / all method calls a bit.
- Should we try to get some commands to move over to using `path` instead of `key` internally?