@jtadmor 2019-03-07
Thanks for the review, Ian.

I'll try and split this PR into a few different ones, so we can get something in that lets all the `Node.get` methods accept paths.

I'm fine with dropping the `byPath` methods from the interface itself. I would still think it might be a good idea to have some utility methods that:
(a) bypass the `resolvePath` lookup when we know we have a path
(b) return a `[node, path]` type signature. Even if we are accepting paths in our methods, if we turn around and look everything up by key, we aren't really utilizing any of the advantages. This could definitely be rolled into a different PR though.