@ianstormtaylor 2019-03-08
@jtadmor for the (a) case there, it won't matter as soon as we remove keys, which we're hoping to do soon, since we won't need the `resolvePath` calls I think? For (b) I agree that sometimes returning `[node, path]` is a nice touch, and I'm not against that at all, I like that solution you came up with.

Thanks!