@jtadmor 2019-03-12
It seems like the one area that I had to overstep to make some changes was the `point` model.

The issue I needed to address is a situation where a range is passed into `normalize` that contains both a `path` and `key`, but where the two are not in alignment. This was passing through tests unnoticed before because we were not using the `path` as much for look-ups, but not that we are, tests were failing.

I think there are a couple approaches we could take. One potentially less intrusive one is to replace the following code by always calling `node.getPath(target.key)`, thus ensuring the path is always valid.

```
    const point = this.merge({
      key: target.key,
      path: path == null ? node.getPath(target.key) : path,
      offset: offset == null ? 0 : Math.min(offset, target.text.length),
    })
})
```

My approach should be roughly equivalent to your suggestion of calling `getNode(path || key)`, except that it also fixes the case where `path` exists but no longer points to a valid node, in which case all we need to do is look it up by `key` and adjust the path. If we don't add that code, we end up hitting the following code:

```
    if (!target) {
      warning(false, "A point's `path` or `key` invalid and was reset!")
      const text = node.getFirstText()
      if (!text) return Point.create()
      const point = this.merge({
        key: text.key,
        offset: 0,
        path: node.getPath(text.key),
      })
      return point
    }
}
```

We don't want to hit this just because the `path` was wrong.