@jtadmor 2019-03-12
Perf (vs master) with current changes to `point` model:

```
  benchmarks
    changes
      ✔ add-mark: 80.81 -> 90.05 ops/sec (11.44% faster)
      ✔ delete-backward: 101.99 -> 104.70 ops/sec (2.66% faster)
      ✔ delete-forward: 102.61 -> 116.56 ops/sec (13.60% faster)
      ✔ insert-fragment-deep-with-problems: 87.06 -> 100.77 ops/sec (15.74% faster)
      ✔ insert-fragment-large: 4.80 -> 4.68 ops/sec (3.00% slower)
      ✔ insert-fragment: 88.58 -> 92.93 ops/sec (4.91% faster)
      ✔ insert-node-by-key: 162.67 -> 190.95 ops/sec (17.38% faster)
      ✔ insert-text-by-key: 199.31 -> 188.30 ops/sec (6.00% slower)
      ✔ insert-text: 84.82 -> 92.63 ops/sec (9.20% faster)
      ✔ normalize-with-problems: 12.37 -> 12.68 ops/sec (2.47% faster)
      ★ normalize:
            user: 100.92 -> 134.35 ops/sec (33.13% faster)
            real: 150.49 -> 210.20 ops/sec (39.67% faster)
      ✔ remove-node-by-key: 106.96 -> 113.18 ops/sec (5.82% faster)
      ✔ set-blocks: 128.69 -> 140.35 ops/sec (9.06% faster)
      ★ split-block:
            user: 52.44 -> 57.76 ops/sec (10.16% faster)
            real: 71.26 -> 96.05 ops/sec (34.80% faster)
    models
      ✔ from-json-big: 14.33 -> 14.53 ops/sec (1.43% faster)
      ✔ from-json: 776.59 -> 912.43 ops/sec (17.49% faster)
      ★ get-active-marks-at-range:
            user: 791.31 -> 8067.54 ops/sec (919.52% faster 🙌 )
            real: 1233.13 -> 16281.77 ops/sec (1220.36% faster 😱 )
      ✔ get-ancestors: 849.76 -> 1109.30 ops/sec (30.54% faster)
      ✔ get-blocks: 13428.36 -> 13943.85 ops/sec (3.84% faster)
      ✔ get-common-ancestor: 1128.40 -> 1136.00 ops/sec (0.67% faster)
      ★ get-descendant:
            user: 922.33 -> 1136.45 ops/sec (23.21% faster)
            real: 1356.25 -> 1808.83 ops/sec (33.37% faster)
      ✔ get-furthest-ancestor: 1707.33 -> 1718.72 ops/sec (0.67% faster)
      ✔ get-inlines: 20340.92 -> 16256.84 ops/sec (25.00% slower)
      ★ get-leaf-blocks-at-range:
            user: 774.97 -> 1730.92 ops/sec (123.35% faster 🙌 )
            real: 1519.80 -> 3207.27 ops/sec (111.03% faster 🙌 )
      ★ get-leaf-inlines-at-range:
            user: 688.52 -> 1025.94 ops/sec (49.01% faster)
            real: 1157.08 -> 1617.37 ops/sec (39.78% faster)
      ✔ get-leaves: 711462.68 -> 671281.78 ops/sec (6.00% slower)
      ✔ get-marks-at-range: 868.16 -> 947.32 ops/sec (9.12% faster)
      ✔ get-marks: 3777.36 -> 3557.86 ops/sec (6.00% slower)
      ★ get-parent:
            user: 3104.61 -> 4775.34 ops/sec (53.81% faster)
            real: 4949.73 -> 5977.02 ops/sec (20.75% faster)
      ✔ get-path: 5547.45 -> 5301.61 ops/sec (5.00% slower)
      ★ get-texts-at-range:
            user: 1889.71 -> 4717.25 ops/sec (149.63% faster 🙌 )
            real: 2469.09 -> 7300.67 ops/sec (195.68% faster 🙌 )
      ✔ get-texts: 18000.77 -> 16905.00 ops/sec (6.00% slower)
      ★ has-node-multiple:
            user: 2522.42 -> 3876.44 ops/sec (53.68% faster)
            real: 3474.16 -> 5067.83 ops/sec (45.87% faster)
      ★ has-node:
            user: 3291.46 -> 4973.10 ops/sec (51.09% faster)
            real: 4571.57 -> 6294.98 ops/sec (37.70% faster)
      ✔ to-json: 3797.19 -> 4113.14 ops/sec (8.32% faster)
    html-serializer
      ✔ deserialize: 186.33 -> 179.87 ops/sec (4.00% slower)
      ✔ serialize: 750.78 -> 622.86 ops/sec (21.00% slower)
    plain-serializer
      ★ deserialize:
            user: 3261.39 -> 4002.36 ops/sec (22.72% faster)
            real: 6153.33 -> 8429.92 ops/sec (37.00% faster)
      ✔ serialize: 7439.56 -> 6969.68 ops/sec (7.00% slower)
    rendering
      ✔ decoration: 66.64 -> 68.55 ops/sec (2.87% faster)
      ✔ large-with-selection: 45.59 -> 46.60 ops/sec (2.22% faster)
      ✔ normal: 77.87 -> 79.86 ops/sec (2.56% faster)
```


Perf with my other proposed change:

```
benchmarks
    changes
      ✔ add-mark: 80.81 -> 86.93 ops/sec (7.57% faster)
      ✔ delete-backward: 101.99 -> 108.76 ops/sec (6.65% faster)
      ★ delete-forward:
            user: 102.61 -> 117.44 ops/sec (14.45% faster)
            real: 153.12 -> 206.45 ops/sec (34.83% faster)
      ✔ insert-fragment-deep-with-problems: 87.06 -> 111.72 ops/sec (28.32% faster)
      ✔ insert-fragment-large: 4.80 -> 4.70 ops/sec (2.00% slower)
      ✔ insert-fragment: 88.58 -> 90.45 ops/sec (2.11% faster)
      ✔ insert-node-by-key: 162.67 -> 166.62 ops/sec (2.42% faster)
      ✔ insert-text-by-key: 199.31 -> 200.73 ops/sec (0.71% faster)
      ✔ insert-text: 84.82 -> 88.31 ops/sec (4.11% faster)
      ✔ normalize-with-problems: 12.37 -> 12.38 ops/sec (0.05% faster)
      ★ normalize:
            user: 100.92 -> 131.60 ops/sec (30.41% faster)
            real: 150.49 -> 207.06 ops/sec (37.59% faster)
      ✔ remove-node-by-key: 106.96 -> 100.75 ops/sec (6.00% slower)
      ✔ set-blocks: 128.69 -> 136.21 ops/sec (5.84% faster)
      ✔ split-block: 52.44 -> 54.12 ops/sec (3.21% faster)
    models
      ✔ from-json-big: 14.33 -> 15.00 ops/sec (4.66% faster)
      ✔ from-json: 776.59 -> 819.72 ops/sec (5.55% faster)
      ✔ get-active-marks-at-range: 791.31 -> 836.93 ops/sec (5.77% faster)
      ✔ get-ancestors: 849.76 -> 1080.67 ops/sec (27.17% faster)
      ✔ get-blocks: 13428.36 -> 14503.10 ops/sec (8.00% faster)
      ✔ get-common-ancestor: 1128.40 -> 1107.27 ops/sec (2.00% slower)
      ✔ get-descendant: 922.33 -> 1088.12 ops/sec (17.97% faster)
      ✔ get-furthest-ancestor: 1707.33 -> 1707.62 ops/sec (0.02% faster)
      ✔ get-inlines: 20340.92 -> 17841.61 ops/sec (14.00% slower)
      ✔ get-leaf-blocks-at-range: 774.97 -> 976.56 ops/sec (26.01% faster)
      ★ get-leaf-inlines-at-range:
            user: 688.52 -> 858.79 ops/sec (24.73% faster)
            real: 1157.08 -> 1691.14 ops/sec (46.16% faster)
      ✔ get-leaves: 711462.68 -> 742397.55 ops/sec (4.35% faster)
      ✔ get-marks-at-range: 868.16 -> 671.51 ops/sec (29.00% slower)
      ✔ get-marks: 3777.36 -> 3012.14 ops/sec (25.00% slower)
      ★ get-parent:
            user: 3104.61 -> 4977.22 ops/sec (60.32% faster)
            real: 4949.73 -> 5811.46 ops/sec (17.41% faster)
      ✔ get-path: 5547.45 -> 5282.14 ops/sec (5.00% slower)
      ✔ get-texts-at-range: 1889.71 -> 2349.50 ops/sec (24.33% faster)
      ✔ get-texts: 18000.77 -> 17436.63 ops/sec (3.00% slower)
      ★ has-node-multiple:
            user: 2522.42 -> 4095.73 ops/sec (62.37% faster)
            real: 3474.16 -> 4900.46 ops/sec (41.05% faster)
      ★ has-node:
            user: 3291.46 -> 4982.40 ops/sec (51.37% faster)
            real: 4571.57 -> 5967.59 ops/sec (30.54% faster)
      ✔ to-json: 3797.19 -> 4451.78 ops/sec (17.24% faster)
    html-serializer
      ✔ deserialize: 186.33 -> 165.02 ops/sec (13.00% slower)
      ✔ serialize: 750.78 -> 658.09 ops/sec (14.00% slower)
    plain-serializer
      ✔ deserialize: 3261.39 -> 3810.37 ops/sec (16.83% faster)
      ✔ serialize: 7439.56 -> 7438.26 ops/sec (0.00% slower)
    rendering
      ✔ decoration: 66.64 -> 64.56 ops/sec (3.00% slower)
      ✔ large-with-selection: 45.59 -> 45.96 ops/sec (0.82% faster)
      ✔ normal: 77.87 -> 77.32 ops/sec (1.00% slower)
```